add shortcuts to link port from current block
Add a shortcut (I) to link current block's first input port to first
output port of the block that was the previous current block. Use case:
one adds a new block below the current on. After that the new block
becomes the current block. This shortcut easily links the two. There's
another shortcut (O) to link output to previous input.